Move get_current_version() script utility function to script_utils
This is so it can be used by other release-related scripts.
